Add two Excel file.
They are the two parts of a big one. I'had to separate it for its size. 
Country and Continent. Pivot table: this is a very simple excel workbook. I created a table with countries and relative continent. It was super useful for creating a hierarchy in the second file "Gerarchia Exercise".